First Upload
Upload of the first draft of ImageSwap puzzle code. The files uploaded are GPIOGame_Master.py and GPIOGame_Slave.py
These are the two scripts for the master and slave raspberry pi to run for the game. Master controls the display and reads inputs. Slave sends signals over wires to master.